Update Management Client to include APIEndpointConfiguration and AIModel
