+++ tests/dotnet/Management.Client.Tests/Clients/Resources/ConfigurationManagementClientTests.cs
@@ -3,11 +3,9 @@ using FoundationaLLM.Client.Management.Interfaces;
 using FoundationaLLM.Common.Constants.Authentication;
 using FoundationaLLM.Common.Constants.ResourceProviders;
 using FoundationaLLM.Common.Models.ResourceProviders;
-using FoundationaLLM.Common.Models.ResourceProviders.AIModel;
 using FoundationaLLM.Common.Models.ResourceProviders.Configuration;
 using FoundationaLLM.Common.Models.ResourceProviders.Vectorization;
 using NSubstitute;
-using System.Collections.Generic;
 
 namespace Management.Client.Tests.Clients.Resources
 {
@@ -202,7 +200,7 @@ namespace Management.Client.Tests.Clients.Resources
 
             // Act & Assert
             var exception = await Assert.ThrowsAsync<Exception>(() => _configurationClient.GetExternalOrchestrationServiceAsync(serviceName));
-            Assert.Equal($"APIEndpointConfiguration '{serviceName}' not found.", exception.Message);
+            Assert.Equal($"ExternalOrchestrationService '{serviceName}' not found.", exception.Message);
         }
 
         [Fact]
@@ -269,40 +267,6 @@ namespace Management.Client.Tests.Clients.Resources
                 ResourceProviderNames.FoundationaLLM_Configuration,
                 $"{ConfigurationResourceTypeNames.APIEndpointConfigurations}/{apiEndpointConfiguration.Name}",
                 apiEndpointConfiguration
-            );
-        }
-
-        [Fact]
-        public async Task UpsertAIModel_ShouldReturnUpsertResult()
-        {
-            // Arrange
-            var aiModel = new AIModelBase
-            {
-                Name = "test-configuration",
-                EndpointObjectId = ""
-            };
-            var expectedUpsertResult = new ResourceProviderUpsertResult
-            {
-                ObjectId = "test-object-id"
-            };
-
-            _mockRestClient.Resources
-                .UpsertResourceAsync(
-                    ResourceProviderNames.FoundationaLLM_AIModel,
-                    $"{AIModelResourceTypeNames.AIModels}/{aiModel.Name}",
-                    aiModel
-                )
-                .Returns(Task.FromResult(expectedUpsertResult));
-
-            // Act
-            var result = await _configurationClient.UpsertAIModel(aiModel);
-
-            // Assert
-            Assert.Equal(expectedUpsertResult, result);
-            await _mockRestClient.Resources.Received(1).UpsertResourceAsync(
-                ResourceProviderNames.FoundationaLLM_AIModel,
-                $"{AIModelResourceTypeNames.AIModels}/{aiModel.Name}",
-                aiModel
             );
         }
     }
